+++ src/hooks/user/usePasswordChangeToggle.ts
@@ -1,0 +1,6 @@
+import { useState } from 'react'
+
+export function usePasswordChangeToggle() {
+    const [showPasswordChange, setShowPasswordChange] = useState(false)
+    return { showPasswordChange, setShowPasswordChange }
+}
